+++ src/snarky.d.ts
@@ -689,6 +689,7 @@ declare const Pickles: {
       publicOutputSize: number;
       storable?: Pickles.Cache;
       overrideWrapDomain?: 0 | 1 | 2;
+      chunks: number;
     }
   ) => {
     provers: MlArray<Pickles.Prover>;